Add files via upload
For some reason the content of a div styled 'display: block' in the CSS stylesheet still shows up when you open the index page. This div should stay hidden until the button for that div is pushed (in this case the button 'about' in the side menu). But apparently, the content of that div is visible,...
Why ow why is this happening? 
Weird part: if you change the order of the three divs for the pages, you will notice that it is always the first div in the order that will show up in index.