Add files via upload
There is a problem with the pageSettings and/or pageAbout / and or pageTutorials div. The divs should remain hidden (display:block) until in Javascript the corresponding function is activated by pushing the corresponding button in the side menu.
Yet, the content of the first div (by order in css), shows up as text in index.php

Also, an editing button might be impossible in this set up.